Only consider name and sorted templates when checking if jobs changed
When determining if a job has changed, we only consider the `name` and
`templates` keys in the job spec. The additional keys are legacy keys
from when agents only supported running one template.

Unfortunately, we cannot remove the legacy keys entirely as they are
still relied upon when running the default errand on any instance group
(i.e. the first specified job). We updated the naming of the variables
to reflect this.

We also removed `spec=` from the release asset templates as this will
cause the templates to be regenerated in a different format when
updating the order of jobs. This is expected behavior if you are
templating the raw spec into your job template.

[#164174731](https://www.pivotaltracker.com/story/show/164174731)

Co-authored-by: Rebecca Putinski <rputinski@pivotal.io>
Co-authored-by: Behrouz Soroushian <bsoroushian@pivotal.io>